# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for Car Catalog App",
        "title": "Car Catalog App API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/cars": {
            "get": {
                "description": "Get All cars with filtration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get All cars with filtration",
                "operationId": "get-all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CarCatalog.Car"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Change record by ID",
                "summary": "Change record by ID",
                "operationId": "change-by-id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "no"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create new car record",
                "summary": "Create new car record",
                "operationId": "add-new",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "no"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete car from list by regNum",
                "summary": "Delete car from list by regNum",
                "operationId": "delete-by-id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "no"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "CarCatalog.Car": {
            "type": "object",
            "properties": {
                "mark": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/CarCatalog.Human"
                },
                "regNum": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "CarCatalog.Human": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        }
    }
}